+++ CHANGELOG.md
@@ -1,10 +1,23 @@
+# 1.84
+- add voice commands for 'fr-FR' and 'it-IT'
+- support protected space for 'fr-FR'
+- support multipe voice commands for the same command, separate wit '|'
+
+# 1.83.x
+- bug fixes
+
 # 1.83
-- Copoilot
-- DOMPurify
+- add service 'Copoilot' ([web.telegram.org](https://web.telegram.org/))
+- support DOMPurify
 
 # 1.82
-- Gemini
-- Trusted Types
+- add service 'Gemini' ([gemini.google.com](https://gemini.google.com/))
+- support Trusted Types
 
- # 1.81
- - initial development
+# 1.81
+- add service 'ChatGPT' ([chatgpt.com](https://chatgpt.com/))
+- add service 'ChatGPT w/o registration' ([seoschmiede.at](https://seoschmiede.at/aitools/chatgpt-tool/))
+- support HTML templates and definitions per service
+
+# 1.x
+- initial development for 'Telegram' ([web.telegram.org](https://web.telegram.org/))